.scripts/requirements.txt
arrow==0.12.1
astroid==1.6.4
backcall==0.1.0
decorator==4.3.0
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
polib==1.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.9.1
python-arango==4.1.0
regex==2018.2.21
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
translate-toolkit==2.3.0
wcwidth==0.1.7
wrapt==1.10.11
